+++ apps/web/app/(base-org)/developers/page.tsx
@@ -1,7 +1,9 @@
-import AnalyticsProvider from 'apps/web/contexts/Analytics';
-import { UseCases } from 'apps/web/src/components/Developers/UseCases';
-import { Hero } from 'apps/web/src/components/Developers/Hero';
 import type { Metadata } from 'next';
+import AnalyticsProvider from 'apps/web/contexts/Analytics';
+import Container from 'apps/web/src/components/base-org/Container';
+import { Hero } from 'apps/web/src/components/Developers/Hero';
+import { UseCases } from 'apps/web/src/components/Developers/UseCases';
+import { Customers } from 'apps/web/src/components/Developers/Customers';
 
 export const metadata: Metadata = {
   metadataBase: new URL('https://base.org'),
@@ -15,10 +17,13 @@ export const metadata: Metadata = {
 export default async function Developers() {
   return (
     <AnalyticsProvider context="developers">
-      <main className="flex min-h-screen w-full flex-col items-center bg-black">
-        <Hero />
-        <UseCases />
-      </main>
+      <Container>
+        <main className="mb-32 flex min-h-screen w-full flex-col items-center bg-black">
+          <Hero />
+          <UseCases />
+          <Customers />
+        </main>
+      </Container>
     </AnalyticsProvider>
   );
 }
